requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.5
django-cors-headers==3.12.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==2.4.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.1.0
